refactoring cli namespace
