+++ ace/system/cli/system.py
@@ -1,12 +1,8 @@
-# vim: ts=4:sw=4:et:cc=120
-
 import base64
 import getpass
 import os
-import sys
-import tempfile
 
-from ace.cli import display_analysis
+
 from ace.constants import ACE_ADMIN_PASSWORD
 from ace.crypto import (
     initialize_encryption_settings,
@@ -16,89 +12,17 @@ from ace.crypto import (
     ENV_CRYPTO_ITERATIONS,
     ENV_CRYPTO_ENCRYPTED_KEY,
 )
-from ace.env import get_base_dir
 from ace.logging import get_logger
-from ace.module.manager import AnalysisModuleManager, CONCURRENCY_MODE_PROCESS, CONCURRENCY_MODE_THREADED
 from ace.system.database import DatabaseACESystem
 from ace.system.database.schema import Base
 from ace.system.default import DefaultACESystem
 from ace.system.threaded import ThreadedACESystem
-from ace.system.remote import RemoteACESystem
 
 
 class CommandLineSystem(DatabaseACESystem, ThreadedACESystem):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.db_url = db_url = "sqlite+aiosqlite://"
-
-
-async def analyze(args):
-
-    if len(args.targets) % 2 != 0:
-        get_logger().error("odd number of arguments (you need pairs of type and value)")
-        return False
-
-    targets = args.targets
-
-    # did we specify targets from stdin?
-    if args.from_stdin:
-        for o_value in sys.stdin:
-            # the type of observables coming in on stdin is also specified on the command line
-            targets.append(args.stdin_type)
-            targets.append(o_value.strip())
-
-    is_local = True
-    uri = get_uri()
-    api_key = get_api_key()
-
-    if uri and api_key:
-        is_local = False
-        system = RemoteACESystem(uri, api_key)
-    else:
-        system = CommandLineSystem()
-        system.storage_root = tempfile.mkdtemp()
-
-    await system.initialize()
-
-    # TODO move this to the system
-    if is_local:
-        await system.create_database()
-
-    manager = AnalysisModuleManager(system, type(system), (system.db_url,), concurrency_mode=CONCURRENCY_MODE_PROCESS)
-    manager.load_modules()
-
-    if not manager.analysis_modules:
-        get_logger().error("no modules loaded")
-        return False
-
-    if is_local:
-        for module in manager.analysis_modules:
-            await system.register_analysis_module_type(module.type)
-
-    root = system.new_root()
-
-    if args.analysis_mode:
-        root.analysis_mode = args.analysis_mode
-
-    index = 0
-    while index < len(args.targets):
-        o_type = args.targets[index]
-        o_value = args.targets[index + 1]
-
-        # TODO if you add a file then add_observable should call add_file
-        if o_type == "file":
-            await root.add_file(o_value)
-        else:
-            root.add_observable(o_type, o_value)
-
-        index += 2
-
-    await root.submit()
-    await manager.run_once()
-
-    root = await system.get_root_analysis(root)
-    display_analysis(root)
-    return True
+        self.db_url = "sqlite+aiosqlite://"
 
 
 async def initialize(args) -> bool:
@@ -119,8 +43,8 @@ async def initialize(args) -> bool:
         encryption_password = getpass.getpass("Enter the ACE administration password:")
 
     if not encryption_password:
-        logging.error(f"missing {ACE_ADMIN_PASSWORD}")
-        sys.exit(1)
+        get_logger().error(f"missing {ACE_ADMIN_PASSWORD}")
+        return False
 
     get_logger().info("initializing database")
     Base.metadata.bind = system.engine
@@ -160,20 +84,3 @@ async def initialize(args) -> bool:
 def initialize_argparse(parser, subparsers):
     init_parser = subparsers.add_parser("initialize", help="Initialize a new ACE system.")
     init_parser.set_defaults(func=initialize)
-
-    analyze_parser = subparsers.add_parser("analyze", help="Analyze given observables.")
-    # options on the RootAnalysis
-    analyze_parser.add_argument("-m", "--analysis-mode", help="Sets the analysis mode of the root.")
-    analyze_parser.add_argument(
-        "--from-stdin",
-        action="store_true",
-        default=False,
-        help="Read observables from standard input. Default observable type is file. Use --stdin-type to change the type.",
-    )
-    analyze_parser.add_argument(
-        "--stdin-type",
-        default="file",
-        help="Specify the observable type when reading observables from stdin. Defaults to file.",
-    )
-    analyze_parser.add_argument("targets", nargs="*", help="One or more pairs of indicator types and values.")
-    analyze_parser.set_defaults(func=analyze)
